// hc_web_angular-Develop/src/app/modules/anestesia-eco-perioperatorio/componentes/interfaces/IdatosPaciente.ts
export interface IdatosPaciente {
  tipoidenpaciente: string;
  identificacionpaciente: string;
  ingreso: number;
  especialidadmedico: string;
  procedimiento: string;
  registromedicorealiza: string;
  nitingreso: string;
  viaingreso: string;
  planingreso: string;
  habitacionpaciente: string;
  posnopos: string;
  cobrable: string;
  usuariocreacion: string;
  programacreacion: string;
  descripcion: string;
  consecutivocita: string;
}